backend/pyproject.toml
[project]
name = "ai-chatbot-backend"
version = "0.1.0"
description = "LangGraph-powered AI chatbot backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # LangChain / LangGraph
    "langchain>=0.3",
    "langgraph>=0.3",
    "langchain-openai>=0.3",
    "langchain-community>=0.3",
    "langgraph-checkpoint-sqlite>=3.0.3",
    # FastAPI + server
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    # Pydantic + settings
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    # Database (async)
    "sqlalchemy>=2.0",
    "aiosqlite>=0.20",       # SQLite async driver (dev)
    # asyncpg>=0.29",        # Uncomment for PostgreSQL in production
    # HTTP / utilities
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "requests>=2.32",
    "typing_extensions>=4.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
